requirements.txt
Pillow
numpy
# Optional later (don’t install yet):
# onnxruntime
# onnxruntime-gpu
# controlnet-aux
